refactor(settings): extract constants to separate file
